+++ src/index.js
@@ -1,4 +1,4 @@
-import { GameStart } from "./game.js";
+import { GameStart } from "./components/game.js";
 
 export default class RacingCarGame {
     constructor() {
