+++ src/input/inputView.js
@@ -1,5 +1,5 @@
-import { makeCarNameIfPossible } from "./carNameCheckor.js";
-import { makeRacingCountIfPossible } from "./countCheckor.js";
+import { makeCarNameIfPossible } from "../components/carNameCheckor.js";
+import { makeRacingCountIfPossible } from "../components/countCheckor.js";
 import { getCarNameInput, getRacingCountInput } from "./inputUtil.js";
 
 export const creatCars = () => {
